requirements.txt
azure-cognitiveservices-speech
azure-identity
pydub
python-dotenv
azure-common
azure-storage-blob